reduce method parameters as suggested by sonar
